+++ jdaviz/core/loaders/importers/spectrum1d/spectrum1d.vue
@@ -33,15 +33,18 @@
     >
     </plugin-auto-label>
     <v-row v-if="data_label_is_prefix">
-        <v-chip
-          v-for="suff in data_label_suffices"
-          outlined
-          label
+        <j-tooltip v-for="(suff, index) in data_label_suffices"
           :key="suff"
-          style="margin: 4px"
-        >
-          {{data_label_value}}{{suff}}
-        </v-chip>
+          :tooltipcontent="data_label_overwrite_by_index[index] ? 'Will overwrite existing entry' : 'New entry'">
+          <v-chip
+            outlined
+            label
+            style="margin: 4px"
+          >
+            <v-icon v-if="data_label_overwrite_by_index[index]" small left color="warning">mdi-file-replace</v-icon>
+            {{data_label_value}}{{suff}}
+          </v-chip>
+        </j-tooltip>
     </v-row>
 
     <plugin-viewer-create-new
@@ -62,20 +65,17 @@
       hint="Select the viewer to use for the new data."
     ></plugin-viewer-create-new>
 
-    <v-row justify="end">
-      <plugin-action-button
-        :spinner="import_spinner"
-        :disabled="import_disabled || extension_selected.length === 0"
-        :results_isolated_to_plugin="false"
-        :api_hints_enabled="api_hints_enabled"
-        @click="import_clicked">
-        {{ api_hints_enabled ?
-          'ldr.load()'
-          :
-          'Import'
-        }}
-      </plugin-action-button>
-    </v-row>
+    <loader-import-button
+      :spinner="import_spinner"
+      :disabled="import_disabled || extension_selected.length === 0"
+      :api_hints_enabled="api_hints_enabled"
+      api_hint="ldr.load()"
+      :data_label_overwrite="data_label_overwrite"
+      :data_label_is_prefix="data_label_is_prefix"
+      :data_label_suffices="data_label_suffices"
+      :data_label_overwrite_by_index="data_label_overwrite_by_index"
+      @click="import_clicked">
+    </loader-import-button>
 
   </v-container>
 </template>